Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    int a,b;
+    float a,b;
     char oper;
     printf("");
     scanf("%c",&oper);
     printf("");
-    scanf("%d %d",&a,&b);
+    scanf("%f %f",&a,&b);
     if(operator=='+')
     printf("%.2f+%.2f=%.2f\n",a,b,a+b);
     else if(operator=='-')
